// packages/cactus-plugin-ledger-connector-tcs-huawei/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface RunTransactionRequest {
    /**
     * 
     * @type {string}
     * @memberof RunTransactionRequest
     */
    chainCodeName?: string;
    /**
     * Function name that was executed.
     * @type {string}
     * @memberof RunTransactionRequest
     */
    functionName?: string;
    /**
     * List of function arguments.
     * @type {Array<string>}
     * @memberof RunTransactionRequest
     */
    functionArgs?: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof RunTransactionRequest
     */
    crossChainReq?: boolean;
}
/**
 * 
 * @export
 * @interface RunTransactionResponse
 */
export interface RunTransactionResponse {
    /**
     * 
     * @type {string}
     * @memberof RunTransactionResponse
     */
    functionOutput: string;
    /**
     * 
     * @type {boolean}
     * @memberof RunTransactionResponse
     */
    success: boolean;
    /**
     * 
     * @type {string}
     * @memberof RunTransactionResponse
     */
    transactionId: string;
}